# app/src/openapi_server/models/new_pet.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class NewPet(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    NewPet - a model defined in OpenAPI

        name: The name of this NewPet.
        tag: The tag of this NewPet [Optional].
    """

    name: str = Field(alias="name")
    tag: Optional[str] = Field(alias="tag", default=None)

NewPet.update_forward_refs()